+++ packages/pwa-kit-runtime/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v2.0.0-dev.2 (Feb 10, 2022)
+
+- Remove default `body-parser` middleware from express server.
+
 ## v2.0.0-dev.2 (Feb 10, 2022)
 ## v2.0.0-dev.1 (Feb 03, 2022)
 ## v2.0.0-dev.1 (Feb 03, 2022)